refactor: ItemTypeService 다형성수정
