+++ src/main/java/com/yoda/accountProject/calendarItem/service/CalendarItemServiceImpl.java
@@ -8,12 +8,11 @@ import com.yoda.accountProject.calendarItem.dto.CalendarItemResponseDto;
 import com.yoda.accountProject.calendarItem.dto.CalendarItemUpdateDto;
 import com.yoda.accountProject.calendarItem.repository.CalendarItemRepository;
 import com.yoda.accountProject.itemType.domain.ItemType;
-import com.yoda.accountProject.itemType.service.ItemTypeServiceImpl;
+import com.yoda.accountProject.itemType.service.impl.ItemTypeServiceImpl;
 import com.yoda.accountProject.system.exception.ExceptionMessage;
 import com.yoda.accountProject.system.exception.calendarItem.CalendarItemNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
